test(core.manager): CopySoBlocTest在不执行时也跳过tearDown的代码

+++ projects/sdk/core/manager/src/androidTest/java/com/tencent/shadow/core/manager/installplugin/CopySoBlocTest.java
@@ -48,6 +48,8 @@ public class CopySoBlocTest {
 
     @After
     public void tearDown() throws Exception {
+        if (!apkFile.exists()) return;
+
         FileUtils.deleteDirectory(soDir);
         copiedTagFile.delete();
     }
